requirements.txt
pandas>=2.3.2
plotly>=6.3.0
streamlit>=1.49.1
psycopg2-binary>=2.9.7
bcrypt>=4.0.1
python-dotenv>=1.0.0
